Image Input field with form integration
